fix(xpu/rocm): address PR #1110 review round 2 - real bugs and defensive fixes
P1 fixes:
- fast_topk_v2: data_ptr<int64_t> -> data_ptr<int32_t> to match CUDA int32 lengths
- ROCm cum_log_probs: gather selected token prob before add (was adding full
  [batch,vocab] to [batch], shape mismatch)

P2 fixes (real bugs):
- ROCm filtered_probs: clone probs_t before in-place top_k/top_p filtering
  to preserve original probabilities for cum_log_probs/output_all_probs
- ROCm multiMergeCopy: replace std::memcpy with hipMemcpyAsync for device
  pointers (std::memcpy on device memory is UB without HMM)

Defensive fixes:
- SDPA batch decode: reshape 1D block_ids to 2D when num_requests>1 to
  prevent cross-request KV contamination
- ROCm multinomial: add degenerate-row protection (argmax fallback for
  all-zero/NaN rows) to prevent crash on extreme top_k+top_p filtering

+++ rtp_llm/models_py/bindings/xpu/RegisterXpuBaseBindings.hpp
@@ -519,8 +519,9 @@ void registerBaseXpuBindings(py::module& rtp_ops_m) {
                           "fast_topk_v2: lengths must be 1-D with length equal to score.size(0)=", score.size(0),
                           ", got shape [", lengths.sizes(), "]");
                       // Validate lengths values are in [0, score.size(-1)].
-                      auto len_cpu = lengths.to(at::kCPU);
-                      auto* len_ptr = len_cpu.data_ptr<int64_t>();
+                      // lengths is int32 (matching CUDA FastTopKParams::lengths).
+                      auto len_cpu = lengths.to(at::kCPU).to(at::kInt);
+                      auto* len_ptr = len_cpu.data_ptr<int32_t>();
                       for (int64_t i = 0; i < len_cpu.size(0); i++) {
                           TORCH_CHECK(len_ptr[i] >= 0 && len_ptr[i] <= score.size(-1),
                               "fast_topk_v2: lengths[", i, "]=", len_ptr[i],
